Improvements done to ecommerce-ecosystem
- Added cache to some data retrieve
- Kafka is working properly now
- Requests to external payment API are asyncronous
- Updated application.properties of cart, catalog and checkout
- Updated build_and_run momentaneously so it can skip tests (>>REMIND TO ROLLBACK THIS<<)

+++ checkout/src/main/java/br/com/fiap/postech/adjt/checkout/dto/CheckoutRequestDTO.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record CheckoutRequestDTO(
         @ValidUUID String consumerId,
-        @JsonProperty("amount") Double amount,
+        @JsonProperty("amount") Integer amount,
         @JsonProperty("currency") String currency,
         @JsonProperty("payment_method") PaymentMethodRequestDTO paymentMethod
 ) {
